+++ test/responds_test.rb
@@ -12,20 +12,4 @@ class RespondsTest < ActiveSupport::TestCase
     assert_not Plain.new.archival?
   end
 
-  ### Deprecation Zone ###
-
-  test "archival class responds correctly to 'is_archival?'" do
-    ActiveSupport::Deprecation.silence do
-      assert     Archival.is_archival?
-      assert_not Plain.is_archival?
-    end
-  end
-
-  test "archival object responds correctly to 'is_archival?'" do
-    ActiveSupport::Deprecation.silence do
-      assert     Archival.new.is_archival?
-      assert_not Plain.new.is_archival?
-    end
-  end
-
 end